avoid circular dependency
